response error json chepeast price fix

+++ Despegar.Core.Business/Flight/Itineraries/CheapestPrice.cs
@@ -7,8 +7,8 @@ namespace Despegar.Core.Business.Flight.Itineraries
 {
     public class CheapestPrice
     {
-        public int savings { get; set; }
-        public int total_price { get; set; }
+        public double savings { get; set; }
+        public double total_price { get; set; }
         public string departure_date { get; set; }
         public string return_date { get; set; }
     }
